osu/configury: add cc fallback compiler
We don't want fabtests to fail in configure
if mpicc is not available.  So, add 'cc' backup
in AC_CHECK_CC.  In the case that mpicc is not available,
cc will be used for completing configuration.  Make'ing
of the osu tests will be disabled if 'cc' cannot be used
to compile the osu MPI tests.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>